requirements.txt
absl-py==1.4.0
altgraph==0.17.3
ansicon==1.89.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
blessed==1.19.1
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
config==0.5.1
construct==2.10.68
construct-typing==0.5.6
contourpy==1.0.5
crypto==1.4.1
cryptography==41.0.1
cycler==0.11.0
decorator==5.1.1
et-xmlfile==1.1.0
filelock==3.10.0
flatbuffers==23.5.26
fonttools==4.37.4
gast==0.4.0
gmssl==3.2.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpustat==1.0.0
grpcio==1.57.0
h5py==3.9.0
hmmlearn==0.2.8
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
jinxed==1.2.0
joblib==1.2.0
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
more-bisect==0.1.1
Naked==0.1.32
numba==0.56.3
numpy==1.23.4
nvidia-ml-py==11.495.46
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.0
pefile==2023.2.7
Pillow==9.2.0
playsound==1.3.0
pooch==1.6.0
protobuf==4.24.2
psutil==5.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyAudioKits==1.0.6
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyparsing==3.0.9
pysmx==0.4.0
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.2
setGPU==0.0.7
shellescape==3.8.1
six==1.16.0
soundfile==0.11.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
timm==0.6.12
torch==1.13.1
torchaudio==0.13.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.12
utils==1.0.1
Wave==0.0.2
wcwidth==0.2.5
Werkzeug==2.3.7
wrapt==1.15.0
xlrd==2.0.1
zipp==3.15.0
